requirements.txt
# ==============================================================
# Poaceae-Stomata-Detection — requirements.txt
# ==============================================================

# ---- Core ML ----
torch>=2.1.0
torchvision>=0.16.0

# ---- YOLO (student & teacher models) ----
ultralytics>=8.1.0

# ---- DINOv3 backbone dependencies ----
transformers
torchmetrics
omegaconf
termcolor
submitit
ftfy
regex

# ---- Scientific computing ----
numpy>=1.24.0
scipy                # Hungarian matching (linear_sum_assignment)
scikit-learn         # PCA for feature-map visualization

# ---- Image processing ----
Pillow>=9.0.0
scikit-image
PyWavelets

# ---- Visualization & logging ----
matplotlib
tqdm

# ---- Data / Config ----
pandas
PyYAML

# ---- Utilities ----
packaging

# ---- Optional: Flash Attention / memory-efficient attention ----
# xformers            # uncomment if using xformers attention in DINOv3